+++ tests/test_checker/test_noqa.py
@@ -163,6 +163,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS441': 1,
     'WPS442': 2,
     'WPS443': 1,
+    'WPS444': 1,
 
     'WPS500': 1,
     'WPS501': 1,